feat: Update getVideoInfoAsync function to support File and Blob objects
This commit modifies the getVideoInfoAsync function to accept File and Blob objects as the sourceFilename parameter in addition to string URIs. This enhancement allows users to pass local video files directly to the function, improving flexibility and convenience. The change also includes an update to the function signature in the code documentation to reflect the new parameter options. See [VideoInfoResult](https://github.com/hirbod/expo-video-metadata/blob/main/src/ExpoVideoMetadata.types.ts#L1) for more information on the returned metadata.