feat: Add support for video location metadata
This commit adds support for extracting the GPS location where a video was recorded. The `location` property is added to the `VideoInfoResult` type and is populated with latitude, longitude, and altitude values if available in the video's metadata. This feature enhances the capabilities of the `getVideoInfoAsync` function and allows users to access location information associated with videos. See [VideoInfoResult](https://github.com/hirbod/expo-video-metadata/blob/main/src/ExpoVideoMetadata.types.ts#L1) for more information on the updated type definition.